Added a RSA(prime number) encryption function
RSA is an incredibly secure asymmetric encryption algorithm relying on calculations and traits of prime numbers for it's encryption scheme. My encryption tool allows you to generate keys, encrypt numbers with a public key, and decrypt an encrypted number with a private key and part of the public key. I have designed it for usability and believe that it is self explanatory . If you have suggestions, please contact me. (I also added another login account)